CUDA: Update to 12.2.2
cuda: Update to 12.2.2

This replaces the dGPU CUDA recipes so that they are bbappend
files that append onto the meta-tegra recipes, using the same
versions but overriding the different SBSA download path and
checksums values, instead of using a different CUDA versions
and having standalone dGPU CUDA recipes.

This allows the dGPU and iGPU to use the same version of CUDA
instead of having to use different versions that are selected
using PREFERRED_VERSION entries.